+++ test/crow/helpers_test.exs
@@ -1,5 +1,0 @@
-defmodule Crow.HelpersTest do
-  use ExUnit.Case, async: true
-
-  doctest Crow.Helpers
-end