+++ CHANGELOG.md
@@ -11,10 +11,14 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Connection worker processes are no longer spawned under a supervisor. This
   also fixes the issue of one connection terminating also terminating other
   connections.
+- Implementing `Crow.Plugin.Name/0` is now mandatory.
 
 ### Fixed
 - Application configuration is now pulled via the `:crow` application key (in
   `Application.get_env` calls) instead of `Crow`, preventing a compiler warning.
+
+### Removed
+- The `Crow.Helpers` plugin.
 
 
 ## v0.1.1 - 2019-04-06